Fix items so they show compendium information in the powers pane
CB exports give the wrong URLs for items.  This change searches
compendium for the correct item and sets the proper URL.

It requires you have an internet connection when importing the character
to the library.  If you don't, they will show up blank, which is still
better than the server error you get now.