When backing up do not prompt for ongoing damage
With this change, you won't be prompted for ongoing damage and power
recharge, and ready status will not be cleared, when backing up.

If you backup twice, then go forward, you still get prompted on the next
turn button.  Fixing that would require keeping track of the state of
everything for each previous turn, which would be more effort than is
worth doing, I think.  Plus we can't tell if you are going back to fix
something, or rewinding time and doing those turns over again.